Move GenericCounter to its own file; update doc comments
